# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Behavior Verification System API
  description: |
    行为验证系统API文档
    
    提供多种验证码类型、用户认证、环境检测、风险分析等功能。
  version: 2.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: https://api.example.com/api/v1
    description: 生产环境
  - url: http://localhost:8080/api/v1
    description: 开发环境
...
paths:
...
  /auth/login:
    post:
      tags:
        - auth
      summary: 用户登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                password:
                  type: string
                captcha_token:
                  type: string
                behavior_data:
                  type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 0
                  message:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      access_token:
                        type: string
                      refresh_token:
                        type: string
                      expires_in:
                        type: integer
                        example: 900
                      user:
                        type: object
                        properties:
                          id:
                            type: integer
                          username:
                            type: string
                          email:
                            type: string
        '401':
          description: 认证失败
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...